Add PostScript tiger demo
This is a port of the famous PostScript tiger from EPS to a self-contained
demo program for the library.  Like many uses of the tiger, it serves as
both a demonstration of the quality of the rendering and as a benchmark.

When run, it will render the tiger to a fresh canvas instance in memory 64
times and report the elapsed time for each trial, with the best time shown
at the end.  It also saves the rendered image to a TGA file, tiger.tga, on
the first trial.

Compiled by GCC 13.2 with -Ofast -march=native -mtune=native on my new
7950X desktop, the peak speed currently seems to be about 70fps.